scripts/train_model.py
"""Training script for eye-tracking models."""

import argparse
import sys
from pathlib import Path

import torch
import yaml

# Add project root to path
sys.path.append(str(Path(__file__).parent.parent))

from src.models.encoder import DummyEncoder
from src.utils.logger import setup_logger


def load_config(config_path: str) -> dict:
    """Load configuration from YAML file."""
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    return config


def train(config: dict):
    """
    Main training function.
    
    Args:
        config: Configuration dictionary
    """
    logger = setup_logger("training")
    logger.info("Starting training...")
    
    # Initialize model
    model_config = config['model']['encoder']
    model = DummyEncoder(
        input_dim=model_config['input_dim'],
        hidden_dim=model_config['hidden_dim'],
        output_dim=model_config['output_dim'],
        num_layers=model_config['num_layers'],
        dropout=model_config['dropout']
    )
    
    logger.info(f"Model initialized: {model.get_model_info()}")
    
    # Training configuration
    train_config = config['training']
    epochs = train_config['epochs']
    lr = train_config['learning_rate']
    
    # Setup optimizer
    optimizer = torch.optim.Adam(model.parameters(), lr=lr)
    
    logger.info(f"Training for {epochs} epochs with learning rate {lr}")
    
    # Placeholder training loop
    for epoch in range(epochs):
        # TODO: Implement actual training logic
        logger.info(f"Epoch {epoch+1}/{epochs}")
    
    # Save model
    model_path = Path(config['logging']['checkpoint_dir']) / 'best_model.pt'
    model_path.parent.mkdir(parents=True, exist_ok=True)
    torch.save(model.state_dict(), model_path)
    logger.info(f"Model saved to {model_path}")


def main():
    """Main entry point."""
    parser = argparse.ArgumentParser(description="Train eye-tracking model")
    parser.add_argument(
        "--config",
        type=str,
        default="configs/config.yaml",
        help="Path to configuration file"
    )
    args = parser.parse_args()
    
    # Load configuration
    config = load_config(args.config)
    
    # Run training
    train(config)


if __name__ == "__main__":
    main()
