requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data processing
scipy>=1.11.0
pillow>=10.0.0

# Logging and monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0
